Added link features common tests
