--- dartsim/worlds/contact.sdf
@@ -1,93 +1,0 @@
-<?xml version="1.0" ?>
-<sdf version="1.6">
-  <world name="default">
-    <model name="ground_plane">
-      <static>true</static>
-      <link name="link">
-        <collision name="collision">
-          <geometry>
-            <plane>
-              <normal>0 0 1</normal>
-              <size>100 100</size>
-            </plane>
-          </geometry>
-        </collision>
-        <visual name="visual">
-          <cast_shadows>false</cast_shadows>
-          <geometry>
-            <plane>
-              <normal>0 0 1</normal>
-              <size>100 100</size>
-            </plane>
-          </geometry>
-        </visual>
-      </link>
-    </model>
-    <model name="sphere">
-      <link name="link0">
-        <pose>0 0 0.5 0 0 0</pose>
-        <collision name="sphere_collision">
-          <geometry>
-            <sphere><radius>1</radius></sphere>
-          </geometry>
-        </collision>
-        <visual name="sphere_visual">
-          <geometry>
-            <sphere><radius>1</radius></sphere>
-          </geometry>
-        </visual>
-        <inertial>
-          <mass>0.1</mass>
-        </inertial>
-      </link>
-      <link name="link1">
-        <pose>0 1 0.5 0 0 0</pose>
-        <collision name="sphere_collision">
-          <geometry>
-            <sphere><radius>1</radius></sphere>
-          </geometry>
-        </collision>
-        <visual name="sphere_visual">
-          <geometry>
-            <sphere><radius>1</radius></sphere>
-          </geometry>
-        </visual>
-        <inertial>
-          <mass>1</mass>
-        </inertial>
-      </link>
-      <link name="link2">
-        <pose>1 0 0.5 0 0 0</pose>
-        <collision name="sphere_collision">
-          <geometry>
-            <sphere><radius>1</radius></sphere>
-          </geometry>
-        </collision>
-        <visual name="sphere_visual">
-          <geometry>
-            <sphere><radius>1</radius></sphere>
-          </geometry>
-        </visual>
-        <inertial>
-          <mass>2</mass>
-        </inertial>
-      </link>
-      <link name="link3">
-        <pose>1 1 0.5 0 0 0</pose>
-        <collision name="sphere_collision">
-          <geometry>
-            <sphere><radius>1</radius></sphere>
-          </geometry>
-        </collision>
-        <visual name="sphere_visual">
-          <geometry>
-            <sphere><radius>1</radius></sphere>
-          </geometry>
-        </visual>
-        <inertial>
-          <mass>3</mass>
-        </inertial>
-      </link>
-    </model>
-  </world>
-</sdf>
