+++ test/common_test/worlds/empty.sdf
@@ -1,0 +1,5 @@
+<?xml version="1.0"?>
+<sdf version="1.6">
+  <world name="empty">
+  </world>
+</sdf>
